Fix integer overflow bug in digraph indexing
Changed digraph indices from u8 to u16 throughout the codebase.

The dictionary had 584 unique digraphs, but u8 can only hold values
0-255. When indices exceeded 255, they wrapped around due to integer
overflow, causing multiple digraphs to incorrectly map to the same
index in the HashMap.

For example, both "fl" (index 126) and "re" (index 382) were mapping
to index 126 after truncation (382 % 256 = 126), causing words like
"are" to be incorrectly validated as "ar-fl".

Changes:
- Word.digraph_indices: Vec<u8> -> Vec<u16>
- Word::with_digraph_indices: HashMap<String, u8> -> HashMap<String, u16>
- Dictionary.root_digraph_to_index: HashMap<String, u8> -> HashMap<String, u16>
- Board::playable_dictionary: HashSet<u8> -> HashSet<u16>

Added debug output to verify digraph mappings are now correct.

+++ src/board.rs
@@ -135,12 +135,12 @@ impl Board {
 
     pub fn playable_dictionary(&self, dictionary: &Dictionary) -> Dictionary {
         // Build a set of usable digraph indices by checking which dictionary digraphs are playable on this board
-        let usable_digraph_indices: HashSet<u8> = dictionary.root_digraph_strings
+        let usable_digraph_indices: HashSet<u16> = dictionary.root_digraph_strings
             .iter()
             .enumerate()
             .filter_map(|(idx, digraph_str)| {
                 if self.digraphs.contains(digraph_str) {
-                    Some(idx as u8)
+                    Some(idx as u16)
                 } else {
                     None
                 }
@@ -152,7 +152,9 @@ impl Board {
             .words
             .iter()
             .filter(|word| {
-                word.digraph_indices.iter().all(|&idx| usable_digraph_indices.contains(&idx))
+                word.digraph_indices.iter().all(|&idx| 
+                    usable_digraph_indices.contains(&idx)
+                )
             })
             .cloned()
             .collect();
